Refactor micro_code mod into folder
