+++ vm/ignite/src/micro_code/binop.rs
@@ -1,86 +1,6 @@
 use crate::{runtime::Runtime, VmError};
 use anyhow::Result;
-use bytecode::{self, BinOp, UnOp, Value};
-
-/// Loads a constant value onto the stack.
-///
-/// # Arguments
-///
-/// * `rt` - The runtime to load the constant onto.
-///
-/// * `val` - The value to load onto the stack.
-///
-/// # Errors
-///
-/// Infallible.
-pub fn ldc(rt: &mut Runtime, val: Value) -> Result<()> {
-    rt.stack.push(val);
-    Ok(())
-}
-
-/// Pops a value off the stack.
-///
-/// # Arguments
-///
-/// * `rt` - The runtime to pop the value off of.
-///
-/// # Errors
-///
-/// If the stack is empty.
-pub fn pop(rt: &mut Runtime) -> Result<()> {
-    rt.stack.pop().ok_or(VmError::StackUnderflow)?;
-    Ok(())
-}
-
-/// Executes a unary operation on the top of the stack.
-/// It pops the value off the top of the stack, applies the
-/// operation, and pushes the result back onto the stack.
-///
-/// # Arguments
-///
-/// * `rt` - The runtime to execute the operation on.
-///
-/// * `op` - The operation to execute.
-///
-/// # Errors
-///
-/// If the stack is empty or the operation is not supported for
-/// the type of the value on the stack.
-pub fn unop(rt: &mut Runtime, op: UnOp) -> Result<()> {
-    let val = rt.stack.pop().ok_or(VmError::StackUnderflow)?;
-
-    match val {
-        Value::Unit => {
-            return Err(VmError::IllegalArgument("unit not supported".to_string()).into())
-        }
-        Value::Int(i) => {
-            let result = match op {
-                UnOp::Neg => Value::Int(-i), // Negation
-                UnOp::Not => Value::Int(!i), // Logical negation
-            };
-            rt.stack.push(result);
-        }
-        Value::Float(f) => {
-            let result = match op {
-                UnOp::Neg => Value::Float(-f),
-                _ => return Err(VmError::IllegalArgument("float not supported".to_string()).into()),
-            };
-            rt.stack.push(result);
-        }
-        Value::Bool(b) => {
-            let result = match op {
-                UnOp::Not => Value::Bool(!b),
-                _ => return Err(VmError::IllegalArgument("bool not supported".to_string()).into()),
-            };
-            rt.stack.push(result);
-        }
-        Value::String(_) => {
-            return Err(VmError::IllegalArgument("string not supported".to_string()).into())
-        }
-    }
-
-    Ok(())
-}
+use bytecode::{BinOp, Value};
 
 /// Executes a binary operation on the top two values of the stack.
 /// It pops the two values off the top of the stack, applies the
@@ -158,133 +78,12 @@ pub fn binop(rt: &mut Runtime, op: BinOp) -> Result<()> {
     Ok(())
 }
 
-/// Jumps to the given program counter if the top of the stack is false.
-///
-/// # Arguments
-///
-/// * `rt` - The runtime to execute the operation on.
-///
-/// * `pc` - The program counter to jump to.
-///
-/// # Errors
-///
-/// If the stack is empty or the top of the stack is not a boolean.
-pub fn jof(rt: &mut Runtime, pc: usize) -> Result<()> {
-    let cond = rt.stack.pop().ok_or(VmError::StackUnderflow)?;
-
-    if let Value::Bool(b) = cond {
-        if !b {
-            rt.pc = pc;
-        }
-
-        Ok(())
-    } else {
-        Err(VmError::IllegalArgument("expected bool".to_string()).into())
-    }
-}
-
-/// Jumps to the given program counter.
-///
-/// # Arguments
-///
-/// * `rt` - The runtime to execute the operation on.
-///
-/// * `pc` - The program counter to jump to.
-///
-/// # Errors
-///
-/// Infallible.
-pub fn goto(rt: &mut Runtime, pc: usize) -> Result<()> {
-    rt.pc = pc;
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use bytecode::{self, Value};
-
-    #[test]
-    fn test_ldc() {
-        let mut rt = Runtime::new(vec![]);
-        ldc(&mut rt, Value::Unit).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::Unit);
-
-        ldc(&mut rt, Value::Int(42)).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::Int(42));
-
-        ldc(&mut rt, Value::Float(42.0)).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::Float(42.0));
-
-        ldc(&mut rt, Value::Bool(true)).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::Bool(true));
-
-        ldc(&mut rt, Value::String("hello world".into())).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::String("hello world".into()));
-    }
-
-    #[test]
-    fn test_pop() {
-        let mut rt = Runtime::new(vec![]);
-        ldc(&mut rt, Value::Unit).unwrap();
-        pop(&mut rt).unwrap();
-        assert_eq!(rt.stack.len(), 0);
-
-        let vals = vec![
-            Value::Unit,
-            Value::Int(42),
-            Value::Float(42.0),
-            Value::Bool(true),
-            Value::String("hello world".into()),
-        ];
-        let val_len = vals.len();
-        let mut rt = Runtime::new(vec![]);
-        for val in vals {
-            ldc(&mut rt, val).unwrap();
-        }
-        for _ in 0..val_len {
-            pop(&mut rt).unwrap();
-        }
-        assert_eq!(rt.stack.len(), 0);
-
-        ldc(&mut rt, Value::String("remember".into())).unwrap();
-        ldc(&mut rt, Value::Unit).unwrap();
-        pop(&mut rt).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::String("remember".into()));
-    }
-
-    #[test]
-    fn test_unop() {
-        let mut rt = Runtime::new(vec![]);
-        ldc(&mut rt, Value::Int(42)).unwrap();
-        unop(&mut rt, UnOp::Neg).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::Int(-42));
-
-        ldc(&mut rt, Value::Float(42.0)).unwrap();
-        unop(&mut rt, UnOp::Neg).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::Float(-42.0));
-
-        ldc(&mut rt, Value::Bool(true)).unwrap();
-        unop(&mut rt, UnOp::Not).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::Bool(false));
-
-        ldc(&mut rt, Value::Bool(false)).unwrap();
-        unop(&mut rt, UnOp::Not).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::Bool(true));
-
-        ldc(&mut rt, Value::Unit).unwrap();
-        let result = unop(&mut rt, UnOp::Not);
-        assert!(result.is_err());
-
-        ldc(&mut rt, Value::String("hello world".into())).unwrap();
-        let result = unop(&mut rt, UnOp::Not);
-        assert!(result.is_err());
-
-        ldc(&mut rt, Value::Int(42)).unwrap();
-        unop(&mut rt, UnOp::Not).unwrap();
-        unop(&mut rt, UnOp::Neg).unwrap();
-        assert_eq!(rt.stack.pop().unwrap(), Value::Int(43));
-    }
+    use crate::micro_code::ldc;
+    use crate::runtime::Runtime;
+    use bytecode::{BinOp, Value};
 
     #[test]
     fn test_binop() {
@@ -388,29 +187,5 @@ mod tests {
         ldc(&mut rt, Value::String(" world".into())).unwrap();
         binop(&mut rt, BinOp::Eq).unwrap();
         assert_eq!(rt.stack.pop().unwrap(), Value::Bool(false));
-    }
-
-    #[test]
-    fn test_jof() {
-        let mut rt = Runtime::new(vec![]);
-        ldc(&mut rt, Value::Bool(false)).unwrap();
-        jof(&mut rt, 123).unwrap();
-        assert_eq!(rt.pc, 123);
-
-        let mut rt = Runtime::new(vec![]);
-        ldc(&mut rt, Value::Bool(true)).unwrap();
-        jof(&mut rt, 42).unwrap();
-        assert_eq!(rt.pc, 0);
-
-        ldc(&mut rt, Value::Unit).unwrap();
-        let result = jof(&mut rt, 42);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_goto() {
-        let mut rt = Runtime::new(vec![]);
-        goto(&mut rt, 123).unwrap();
-        assert_eq!(rt.pc, 123);
     }
 }
