+++ vm/ignite/src/micro_code/goto.rs
@@ -1,0 +1,31 @@
+use crate::runtime::Runtime;
+use anyhow::Result;
+
+/// Jumps to the given program counter.
+///
+/// # Arguments
+///
+/// * `rt` - The runtime to execute the operation on.
+///
+/// * `pc` - The program counter to jump to.
+///
+/// # Errors
+///
+/// Infallible.
+pub fn goto(rt: &mut Runtime, pc: usize) -> Result<()> {
+    rt.pc = pc;
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::runtime::Runtime;
+
+    #[test]
+    fn test_goto() {
+        let mut rt = Runtime::new(vec![]);
+        goto(&mut rt, 123).unwrap();
+        assert_eq!(rt.pc, 123);
+    }
+}
