+++ vm/ignite/src/micro_code/mod.rs
@@ -1,0 +1,13 @@
+pub use binop::binop;
+pub use goto::goto;
+pub use jof::jof;
+pub use ldc::ldc;
+pub use pop::pop;
+pub use unop::unop;
+
+mod binop;
+mod goto;
+mod jof;
+mod ldc;
+mod pop;
+mod unop;
